// src/types/user.ts
export interface User {
    id: number;
    name: string;
    username: string;
    email: string;
    age: number;
    address: Address;
    phone: string;
    website: string;
    company: Company;
}
export interface Address {
    street: string;
    suite: string;
    city: string;
    zipcode: string;
    geo: Geo;
}
export interface Geo {
    lat: string;
    lng: string;
}
export interface Company {
    name: string;
    catchPhrase: string;
    bs: string;
}
export interface ReducedUser {
    id?: string;
    username: string;
    address: Address;
    age: number;
    companyname: string;
    restore?: boolean;
}